Project1 submission
The overall goal is to examine how an individual household energy usage
varies over a 2-day period in February, 2007. The task is to
reconstruct the graphic plots in the “Readme.”

The data are measurements of electric power consumption in one
household with a one-minute sampling rate over a period of almost 4
years.

Retrieved from
https://archive.ics.uci.edu/ml/datasets/Individual+household+electric+po
wer+consumption